虚拟人格/rag-engine/requirements.txt
flask>=2.0.0
flask-cors>=3.0.0
sentence-transformers>=2.2.0
faiss-cpu>=1.7.0
numpy>=1.21.0
scikit-learn>=1.0.0
requests>=2.28.0